LayoutTests/webaudio/oscillatornode-legacy-api-expected.txt

Testing that the legacy OscillatorNode API is not exposed on the unprefixed OscillatorNode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS OscillatorNode.prototype.playbackState is undefined.
PASS OscillatorNode.UNSCHEDULED_STATE is 0
PASS OscillatorNode.SCHEDULED_STATE is 1
PASS OscillatorNode.PLAYING_STATE is 2
PASS OscillatorNode.FINISHED_STATE is 3
PASS legacyOscillator.playbackState is defined.
PASS legacyOscillator.UNSCHEDULED_STATE is 0
PASS legacyOscillator.SCHEDULED_STATE is 1
PASS legacyOscillator.PLAYING_STATE is 2
PASS legacyOscillator.FINISHED_STATE is 3
PASS successfullyParsed is true

TEST COMPLETE
